backend/requirements.txt
git+https://github.com/weaviate/weaviate-recommend-python-client.git
python-dotenv
tqdm
wasabi
uvicorn
fastapi